Add replies on rehearse commands
Makes the rehearse plugin reply with comments on a PR in response when a
user uses /rehearse, so that user knows what actually happened after
they added their comment.

In case of `/rehearse ?` a list of all jobs for which a rehearsal can be
triggered is added as comment.

In all other cases a list of jobs for which a rehearsal was triggered is
added as comment.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>