Add a custom dependabot for our projects
The github dependabot is very nice but it has the limitation that it can
not run custom cleanup and bump commands before creating a PR.

This new robot mimics the dependabot and queries the CVE API which
dependabot uses to detect CVEs.

The advantage is, that after running the bot, we can run follow-up
commands in our prow-jobs.

For example:

```
bazel run //robots/cmd/dependabot:dependabot -- --github-token-path /path/to/token --repo kubevirt --org kubevirt --repo-dir /full/path/to/repo
make -C /full/path/to/repo deps-sync
```

Signed-off-by: Roman Mohr <rmohr@google.com>