Move remaining prow jobs to the new workloads cluster (#2781)
The major CI jobs have been running on the new workloads cluster for the
past couple of weeks. This change moves the remaining prow jobs from the
old cluster to the new workloads cluster.

As the old cluster is due to be decommissioned - these jobs have to be
moved over.

Signed-off-by: Brian Carey <bcarey@redhat.com>