Remove files for bootstrapping kubernetes on old workloads clusters
Remove the ansible files that were used for bootstrapping and installing
kubernetes to the old workloads cluster and the old ARM cluster. The ARM
cluster no longer exists and the old prow workloads cluster is due to be
removed soon. The automated jobs that use these files were removed in a
previous PR[1]

[1] https://github.com/kubevirt/project-infra/pull/2792

Signed-off-by: Brian Carey <bcarey@redhat.com>