Update prow images, patch prow deployments to remove unnecessary mounts (#2786)
* Run hack/bump-prow.sh

Signed-off-by: kubevirt-bot <kubevirtbot@redhat.com>

* Patch prow deployments to remove unnecessary mounts

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Change order of remove ops to backwards

kustomize seems to operate on a one by one basis, since if you have a
couple of remove ops in ascending order it doesn't find the later
elements. Therefore we change the removals to be in backwards order aka
from last to first entry.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Move gcs service account addition to the end

We move the addition of arg, mount and volume to the end of the
patch to keep the delete indices consistent. This way it's easier to
follow.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: kubevirt-bot <kubevirtbot@redhat.com>
Signed-off-by: Daniel Hiller <dhiller@redhat.com>
Co-authored-by: kubevirt-bot <kubevirtbot@redhat.com>