Remove old workload clusters deployment files (#2796)
* Remove files for bootstrapping kubernetes on old workloads clusters

Remove the ansible files that were used for bootstrapping and installing
kubernetes to the old workloads cluster and the old ARM cluster. The ARM
cluster no longer exists and the old prow workloads cluster is due to be
removed soon. The automated jobs that use these files were removed in a
previous PR[1]

[1] https://github.com/kubevirt/project-infra/pull/2792

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Remove deployment files for prow components on old cluster

The old workloads cluster is no longer taking any prow jobs and will be
removed soon. Deploying updates to the prow components in this cluster
is no longer required so these files can be removed.

These are deployed to the new cluster as part of this PR [1]

[1] https://github.com/kubevirt/project-infra/pull/2664

Signed-off-by: Brian Carey <bcarey@redhat.com>

---------

Signed-off-by: Brian Carey <bcarey@redhat.com>