Introduce flake stats aggregator (#2833)
* Prototype of flake stats aggregator

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix detection of quarantined tests over report period

Since an addition or removal of a quarantine label in the test name can
lead to additional or less spaces, we strip more than one space to one.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update heat map coloring, add overall section

We now color the cards according to their share of failures from the
overall number of failures.

We also aggregate totals for all tests, per day and per lane.

We add a link to testgrid for the lane cards.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Remove Irrelevant values from aggregates

We remove display of all other aggregated values besides sum and share
of failures to avoid the burden of overwhelming the reader with too
much data.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update bazel files

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add README

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add periodic job

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add test for normalization

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Simplify regex for spaces

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Refine html template

* add link to README
* remove empty cells in filter table
* improve showing number of tests shown

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>