test-label-analyzer: add arg to use filter-file as rule input method (#2889)
* Add -filter-tests-file to consume quarantined_tests.json

Since we want to see which tests are actually being quarantined if we
add them to quarantined_tests.json we enable using this file as an
input.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add MatchingPath to html data carrier

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Refactor: remove obsolete GitBlameLines from html data

Since it's already contained in MatchingPath. Also fix tests,
fix bazel files, extract git.go from stats.go

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Refactor: extract git commands into own package

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add rule age information to filter file configs

In order to display wo added a rule to a filter file (like
quarantined_test.json) we parse the git blame info and attach that to
the item matching the id. We also change the html template to show
that information.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add Hits to LabelCategory to track how many matches occurred

Besides looking at which tests are actually targeted by a configuration
we also want to know whether rules in a configuration are ineffective,
thus we count the number of matches per LabelCategory in the Hits
attribute. This helps us in refining the rule set to be most effective.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add README section for using input file

Also fix some json field names.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update bazel config

Also remove BUILD.bazel from testdata, since it prohibits adding
testdata files to the test sandbox.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Rename go file in testdata to ginkgo

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Load expected git blame info on the fly

Also move Makefile to use bazel instead of go, since the main presubmit
is using bazel.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Try to fix bazel sandbox behaviour for git

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Install ginkgo

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add test data for creating test output

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add filtering of stats input file

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Final cleanup

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Create test-label-analyzer image

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Create test-label-analyzer image jobs

Create presubmit to check image build words, and create publish job that
pushes the image to quay.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update jobs to use test-label-analyzer image

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Switch away from dhiller fork for image build

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Address review comments

* extract file mode constant
* replace %v with %w to use error unwrapping
* simplify multiple ifs with switch case

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>