[performance] Increase performance job memory (#3023)
PR #10110 introduced an additional container
in each virt-launcher pod to stream out the
logs from the serial console of the guest OS.
This additional container is running a custom
tail binary named virt-tail and so it also
requires resources.
The increased amount of requested resources
(35M of memory for each virt-launcher pod)
was making the performance test lane failing
and so the number of VMs involved in the test was
squezed from 100 to 90 but due to that
the results are now not directly comparable.

In order to be able to revert back to 100
test VMs we need 35M*100=3500M more,
let's round to 4Gi more.

Then
https://github.com/kubevirt/kubevirt/pull/10566
(once merged), will reserve 100Mi more as overhead
for guaranteed vmi and all the vmis in the density
tests are configured as guaranteed ones so we
will need 100Mi*100=10000Mi more.

So let's reconfigure the performance lanes:
memory: 56Gi -> 70Gi

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>