flake stats report - add periodic runs, improve most recent failures sort order (#3019)
* Remove filtering of periodic job results

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Sort by using more recent failures as comparison

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Refactor: reduce complexity by extracting methods

Also fix error wrt testgrid urls and periodic jobs.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add --filter-periodic-job-run-results flag

Using the above flag we can select whether we want to include the
periodic results or not. Default is to include the results.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Remove share value in table, use as title instead

Since the share is also displayed via color coding we remove the table
row holding it, and instead use a title attribute on the table so that
users can still see it.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update screenshots and README

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>