release-tool: Perform more thorough error checking
Things could go wrong when creating the release notes file or
the git tag, and we shouldn't silently ignore such failures.

We have to skip generation of release notes in the test suite
because of this change, since otherwise we will get error

  unable to generate release notes because no previous tag detected

as there is no real git repository in the mocked environment.
We were already skipping this part in practice, it's just that
until now we also generated an unreported error as part of the
attempt.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>