release-tool: Introduce --name argument
The tool is supposed to be largely project-agnostic. In reality,
it's currently only used for KubeVirt, as evidenced by the fact
that the name of the project is hardcoded in at least one place.

Since we're going to use the project's name in more places in
the future, get a bit closer to the intended scenario by asking
for it to be provided as a command line argument.

As an exception, detect the project name automatically for the
main KubeVirt repository. This makes it possible to keep calling
the tool without providing the extra argument.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>