release-tool: Calculate number of contributors correctly
The output of 'git shortlog' always ends with a newline, so
calling strings.Split("\n") on it will result in a stray empty
string finding its way into the output. We also skip the entry
for kubevirt-bot late, right before we would print it.

As a consequence, we've always printed out a number of
contributors that's higher (by exactly two) than the number of
items in the contributors list.

Filter out the list early instead, avoiding the problem.

An empty line is added to the top of the "Additional Resources"
text to compensate for the one that is no longer printed at the
end of the contributors list.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>