Update prow-deploy presubmit to use kubevirtci cluster providers (#3061)
* Update prow-deploy presubmit to use kubevirtci cluster providers

The current implementation is quite complicated with two levels of
podman in podman and a kind cluster deployment inside of these layers.
There is a large number of zombie processes created in the current test
pod.

This change simplifies it by testing the deployment of prow services
against a kubevirtci cluster removing the need for molecule and kind.
The issue with zombie processes does not occur with this approach.

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Remove molecule dir as no longer used

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Update prow-deploy README to include new testing method

Also removes references to the old molecule implementation

Signed-off-by: Brian Carey <bcarey@redhat.com>

---------

Signed-off-by: Brian Carey <bcarey@redhat.com>