plugins: Introduce test-subset plugin
Allows to run specific labels on a lane
/test-subset <job-name> <labels>

It will create a job, inject the label to its env var,
and give it a unique suffix, so it won't
affect the official jobs.
Opposed to focus (which is also not encouraged, now that we have labels),
the job won't fail if all lanes pass, which reduce emails,
comments, and the need to manually look on the job results on both cases.

Signed-off-by: Or Shoval <oshoval@redhat.com>