fix, botreview: handle deleted files in kubevirtci-bump, improve commenting  (#3124)
* fix, botreview, strategy: ignore unrelated files in kubevirtci-bump

Previously all files were added as unwanted whenever a review was
placed, even if there we no files related to kubevirtci-bump at all in
the patch set.

Now the review for kubevirtci-bump is only attached if there are
relevant files in the PR.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* botreview, test: make AttackReviewComment testable

Introduce interfaces for faking input and output, create some basic test
cases to fix behavior.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* botreview, comments: create details section

Move the details of the review into a Markdown details section, so that
they do not overflow the comment for the review.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* botreview, comment: convert file list into md bullets

Reduce list of files (in case of a very long comment) into a
real bullet list.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* botreview, kubevirtci-bump: handle deleted files case

If files are deleted inside the provider bump, this is still ok, so
handle file deletions as regular changes.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix, bazel, botreview: update build file

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix, botreview: fix comment formatting

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix, botreview, image-update: fix regex

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>