kubevirt,presubmits: make 1.26 lanes optional (#3176)
Following the introduction of the phased plugin, required jobs that are
manually triggered are allowed[1] - this leaves 1.26 jobs as required
but not triggered on kubevirt PRs which is preventing merge queue PRs
from being merged.

Updating these lanes to be explicitly optional so that they no longer
appeared as required.

[1] https://github.com/kubevirt/project-infra/blob/0c8c659783c24fe4b5b939455169fb6535bf052a/github/ci/prow-deploy/kustom/base/configs/current/config/config.yaml#L274

Signed-off-by: Brian Carey <bcarey@redhat.com>