prow-deploy, fix: Role and RoleBinding for statusreconciler (#3180)
Role and RoleBinding for statusreconciler are only present in
kubevirt-prow, but prowjob objects are created in kubevirt-prow-jobs ns.

We therefore add the missing things for test-pods ns, which is then
changed to kubevirt-prow-jobs during deployment kustomization.

Aside from that we remove an admin account from the setup, since that
account is inactive for quite a while now.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>