Update test-infra dependency to make use of new prowjob config fields
As part of the phased plugin work we are introducing a new field
`run_before_merge` to some of our prowjobs - this is not available in
the version of test-infra that is currently in place and as a result the
automation removes this field[1] from our configs.

Update test-infra to the latest version in order to allow using this
field.

[1] https://github.com/kubevirt/project-infra/pull/3198#discussion_r1461650975

Signed-off-by: Brian Carey <bcarey@redhat.com>