kubevirt,presubmits: increase memory requests for prow-control-plane jobs (#3195)
When there is a large number of jobs kicked off at once, we can see that
the control-plane cluster gets overwhelmed and jobs fail early to start the
podman socket[1].

Doubling the memory request will reduce the number of running jobs by
half and should help with this problem.

These jobs are generally very short so this should not really impact
feedback time.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/11041/pull-kubevirt-verify-go-mod/1748262393516920832

Signed-off-by: Brian Carey <bcarey@redhat.com>