kubevirt,presubmit: remove 1.25 FIPS test lane
This lane is running against an unsupported kubernetes version for
the kubevirt main branch.

There is no kubevirtci cluster provider available to even run this lane
against.

It is causing some of our automation to fail[1] as old unsupported
versions are not expected.

Removing this presubmit - a FIPS job can be added again using a more
recent kubernetes version if its needed.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/logs/periodic-kubevirtci-cluster-unsupported-version-remover/1753079938228948992#1:build-log.txt%3A67

Signed-off-by: Brian Carey <bcarey@redhat.com>