kubevirt-presubmits: Make the lint target job required (#3219)
The external linter on the kubevirt/kubevirt project has been stable for
at least a year now.

The linter protects the areas of code it covers well and it has proven
useful to keep the codebase clean and reduce potential bugs.

Lately the nogo embedded linters have partially been disabled due to
some incompatibility when the Go version and bazel have been upgraded.
Therefore, the external linter is also a good backup to keep around.

As of now, the golangci-lint meta linter is running 31 linters.

Signed-off-by: Edward Haas <edwardh@redhat.com>