Add help-wanted to restricted_labels config
In oder to be usable with /label command, the labels
`good-first-issue` and `help-wanted` have to be added to the
restricted_labels configuration inside the prow configuration file for
plugins.

This change adds `help-wanted` to be usable by `kubevirtorg-label`
GitHub team (defined in orgs.yaml), and also adds `help-wanted` to
`kubevirtkubevirt-label` and `kubevirtproject-infra-label` GitHub teams.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>