referee, prow, plugin: add PR labels to evaluation
We now also consider whether a previous hold was already issued before
by the bot user, and if so, we do not place another hold again.
The hold is not placed even if it is not present any more.
This covers the case where the hold was placed once by the botuser and
then was taken back by someone else.

Note: updated robots/cmd/retests-to-merge to fetch PRs on itself.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>