rehearse, plugin, prow: set commit id from target
When we rehearse a PR for a repo different from project-infra, we need
to set correct git commit information inside the `refs:` section of the
presubmit data, otherwise the rehearse will fail if the rehearsed job
relies on that information.

Rehearsal is done basically by creating a new prowjob from the
definition, where a pull request is the source of the git data that
refers to the changes we are trying to test when creating presubmits.

Now, since we want to test whether our changed presubmit works, we have
to create an artificial pull request data object that has the data
pointing to the target repository that the presubmit is for.

The previous solution simplified this by just reusing the pull request
data of the project-infra pull request, which obviously points to the
wrong repository if the presubmit that is changed is targeting i.e. the
kubevirt/kubevirt repository.
It solved the missing target repository by simply adding an
`extra_ref` pointing at the target repository.

However, that didn't work in all cases, i.e. if the target repo uses
information from the presubmit, like i.e. environment variables injected
from prow.

One example is automation/test.sh from kubevirt/kubevirt, which relies on
PULL_BASE_SHA.

The new solution uses an artificial PR instance per branch, which copies
the original data to preserve the pointers to the original PR, while
changing the ref data to the target repository. The extra_ref can be
removed, since it's no more necessary.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>