referee, prow, plugin: add plugin that holds PRs on too many retest comments  (#3251)
* referee, prow, plugin: initial prototype

This commit provides a simple tool that counts the number of retest
comments after the latest push to the given PR.

Also it provides the server implementation for the prow external plugin.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: add bazel build files

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: add files for publishing image

Add Makefile, Containerfile and job definition for publishing the plugin
image.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: add service and deployment for k8s

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: enable in prow plugin config

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: test event handling

Extract interfaces to be able to mock out graphql client and github
client dependencies. Then create a couple of test cases to make sure
that the botuser doesn't handle it's own comments.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: ping team on hold comment

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: leave PR alone if hold placed

In case a PR already has received a hold comment we don't handle it again.
The intent is that the author and a team has already been pinged on the
PR, and we leave it up to them to handle this now.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: fix review issues

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, prow, plugin: add PR labels to evaluation

We now also consider whether a previous hold was already issued before
by the bot user, and if so, we do not place another hold again.
The hold is not placed even if it is not present any more.
This covers the case where the hold was placed once by the botuser and
then was taken back by someone else.

Note: updated robots/cmd/retests-to-merge to fetch PRs on itself.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* prow, plugin, referee: add unit tests for labels

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* robots, retests-to-merge: add README, refine output

Refines the output of the cli tool to sort the output starting with
highest number of retest PRs first. Adds README and updates main README
to contain entry for tool.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* prow, plugin, referee: update wording in comment

Add references to k/kubevirt CONTRIBUTING and
getting-started#testing.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>