bootstrap,s390x: Add s390x support to multiarch bootstrap build (#3197)
* bootstrap,s390x: Add s390x support to multiarch bootstrap build

There is an effort to enable support for s390x for KubeVirt[1].

This adds an s390x bootstrap image to our multiarch publish flow.

This bootstrap image does not include bazel as it will have to rely on
the s390x builder image[2] for bazel.

This image will allow us to run native s390x builds and unit tests on a
s390x cluster[3].

[1] https://github.com/kubevirt/kubevirt/pull/10490
[2] https://github.com/kubevirt/kubevirt/pull/11097
[3] https://github.com/kubevirt/project-infra/issues/3140

Signed-off-by: Brian Carey <bcarey@redhat.com>

* bootstrap: Remove older version of Bazel from bootstrap

Signed-off-by: Brian Carey <bcarey@redhat.com>

---------

Signed-off-by: Brian Carey <bcarey@redhat.com>