Revert "bootstrap: Set image proxy at the start of jobs (#3291)" (#3296)
This reverts commit 471d8c9bc9ce8ff64cbd5ae94a716273582a6ee5.

Prow jobs had issues downloading dependencies when this change was
applied to the job configs.

Bazel uses the same environment variables for proxy[1] so there may be a
an issue there causing the image proxy cache getting overloaded.

Reverting this change so that it doesn't get added again through another
bump of prow images

[1] https://bazel.build/external/advanced#using_proxies

Signed-off-by: Brian Carey <bcarey@redhat.com>