kubevirtci,publish: Publish kubevirtci using podman
Previously kubevirtci providers that were published with podman used to
break `make cluster-up` for docker users. This was due to an empty
`/run/.containerenv` being added to the image when published with
podman. This broke some IF statements that setup the iptables rules to
communicate with the VM.

This should be fixed when
https://github.com/kubevirt/kubevirtci/pull/1149 is merged.

Signed-off-by: Brian Carey <bcarey@redhat.com>