Fix common-instancetypes-periodics (#3325)
* common-instancetypes-periodics: Use absolute askpass path

This was removed by mistake in af4883765 while trying to clean up the
jobs. However it is still required as suggested by dceff08aaa24.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

* common-instancetypes-periodics: Fix main latest version logic

The /releases/latest API provided by GitHub does not list pre-releases.
We actually want to pick these up for early testing on main so switch
over to the full /releases API to fetch these.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

---------

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>