WIP prow, autobump: add config files
gcr.io/k8s-prow/generic-autobumper needs configuration files per update.
We create two, one for the prow components and another one for the jobs.

Also we modify the bump script so that it runs the autobumper as a pod.
For that we need to add the podman in container preset.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>