prow, autobump: add config files, modify bump script (#3364)
* prow, autobump: add config files, modify bump script

gcr.io/k8s-prow/generic-autobumper needs configuration files per update.
We create two, one for the prow components and another one for the jobs.

Also we modify the bump script so that it runs the autobumper as a pod.
For that we need to add the podman in container preset.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* prow, autobump: make /etc/github the default

Also use a tagged version of the autobump. Therefore we need to add the
script to the extraFiles section in the config, so that it is included
in the bumping.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>