kubevirt,releases: Add change to mdev module name in vgpu release presubmits (#3377)
These vgpu e2e lanes are failing due to missing a kernel module vfio_mdev[1] after the workloads cluster upgrade.

The mdev kernel module is already loaded on the nodes as the nvidia operator is installed.

This updates the vgpu providers on the release branches to laod the mdev
kernel module instead of vfio_mdev.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/logs/periodic-kubevirt-e2e-kind-1.27-vgpu/1784918657499926528#1:build-log.txt%3A310

Signed-off-by: Brian Carey <bcarey@redhat.com>