kubevirt,vgpu: Fix antiaffinity rules for vgpu lanes to prevent collisions (#3381)
I have noticed multiple vgpu e2e lanes running on the one node at once
which causes the test lane to fail as these require access to the GPU on
the node.

This fixes the antiaffinity rules to avoid these collisions.

The SRIOV lanes have similar antiaffinity rules[1] and do not collide on
the SRIOV nodes.

[1] https://github.com/kubevirt/project-infra/blob/5756c30ab7cea66a9aa4334d92b06eb7a2e0a00c/github/ci/prow-deploy/files/jobs/kubevirt/kubevirt/kubevirt-presubmits.yaml#L204C1-L211C25

Signed-off-by: Brian Carey <bcarey@redhat.com>