kubevirt,presubmits: Add sig-compute-serial e2e test lane (#3362)
This sig-compute-serial test lane will be responsible for running all of
the sig-compute serial tests.

The other sig-compute lanes are updated to just run the parallel tests.

This change requires the following PR in kubevirt/kubevirt:
https://github.com/kubevirt/kubevirt/pull/11740

The sig-compute lane is by far the longest running lane - over 3.5 hours
on average[1]. Splitting this lane will give quicker CI feedback to contributors

[1] https://grafana.ci.kubevirt.io/d/gY4hNc5Iz/kubevirt-kubevirt-e2e-test-runtimes?orgId=1&viewPanel=2&from=now-90d&to=now

Signed-off-by: Brian Carey <bcarey@redhat.com>