prow, rehearse: add ok-to-rehearse label check (#3405)
* prow, rehearse: add ok-to-rehearse label check

There are situations where we want to allow rehearsal for users that
otherwise would not be allowed to rehearse (not an approver in the
project).

Therefore we enable rehearsals for pull requests where the label
`ok-to-rehearse` is present.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* prow, rehearse: add ok-to-rehearse label to config

We add the label to the org label configuration, so that people can
apply it to pull requests.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>