sigs, sync: update GitHub team configurations for sigs
Updates the given Peribolos configuration file with teams per sig.
For each sig alias group encountered in the OWNERS_ALIASES it
creates a new team if it doesn't exist, adding the members of
the alias group to the team members section.
If the team does exist, it merges the GitHub handles from the alias
group with the existing members of the team.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>