prow, referee: add retest metrics (#3380)
* prow, referee: add metrics

We add a metrics package that encapsulates the metrics capturing for us,
which we use to abstract away the metrics collection inside the server.

We also add the calling of the metrics funcs to the retest logic.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* prow, referee, metrics: hook up metrics handler to server

We hook the metrics handler to the http server on the default metrics
route.

We create tests for the metrics that are emitted by the handler.

Aside of that we create a basic README.md file that explains the
motivation of the plugin and holds docs for generated metrics.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* prow, referee: initialize retest gauges for repositories

Create a flag to give repository identifiers that the gauges
for pull requests are initialized with. Per repository all
pull requests with labels `lgtm` and `approved` are fetched and for
these the number of retests added to the org/repo gauge.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* referee, metrics, tests: use Eventually for http client call

Due to sporadically happening dial failures the unit tests around
metrics were flaky. By using Eventually we retry the call whenever it
fails. This has shown to not fail in over 10000 consecutive runs
using

```bash
ginkgo run --until-it-fails ./external-plugins/referee/metrics/...
```

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>