shared-images: Wait for currently running jobs to complete before deleting old images
After a bump-kubevirtci PR is merged - the shared images controller will
pull the new images and delete the old images however this can cause
problems for the currently running e2e jobs as they are running with the
old image.

These jobs fail at the very end when trying to clean up the kubevirtci
containers because the old image layers are no longer available.

Adding a delay before cleaning up the old images to allow these running
jobs to complete should resolve this.

Signed-off-by: Brian Carey <bcarey@redhat.com>