periodic,docker-proxy: Add check for docker mirror proxy CA certificate expiration
There was an recent issue where the CA certificate for the
docker-mirror-proxy image cache expired causing the majority of the jobs
that rely on the cache to pull images to fail[1].

This can have a big impact on CI and our ability to merge PRs.

This periodic check will start failing when the CA certificate will
expire in the next 90 days. This should give us enough warning to plan
for renewing this certificate

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_project-infra/3451/pull-project-infra-prow-deploy-test/1797955833284268032#1:build-log.txt%3A14

Signed-off-by: Brian Carey <bcarey@redhat.com>