kubevirt,presubmit: Enable vgpu presubmits and set to required (#3484)
The vgpu presubmit lanes were disabled following the upgrade of the
cluster which caused these lanes to fail constantly[1]

These failures were due to an issue with the version of the vgpu driver
installed on the nodes. Updating to a later version of the driver has
resolved this issue and the periodic vgpu lane is passing
consistently[2].

The vgpu presubmit test lanes can be reintroduced and set as required.

[1] https://github.com/kubevirt/project-infra/pull/3408
[2] https://prow.ci.kubevirt.io/job-history/gs/kubevirt-prow/logs/periodic-kubevirt-e2e-kind-1.27-vgpu

Signed-off-by: Brian Carey <bcarey@redhat.com>