bootstrap: Remove cgroups: disabled from containers.conf (#3515)
When the workloads cluster was upgraded to cgroups v2, an issue with the
vgpu lanes appeared[1]. After some investigation, we suspect that it is
due to cgroups being disabled in podman in container setup.

Removing this option does not cause any impact to other workloads so it
is not required.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/logs/periodic-kubevirt-e2e-kind-1.27-vgpu/1803735246215057408

Signed-off-by: Brian Carey <bcarey@redhat.com>