bootstrap: leave short-name-mode as default enforcing
Setting short-name-mode to permissive was a temporary workaround for
scripts which hadn't been updated to use full image names.

Setting short-name-mode to enforcing forces scripts to be updated with
full image names which is best practice and clarifies which image is
being used.

Any jobs failing following this update should be updated to use the full
image name

Signed-off-by: Brian Carey <bcarey@redhat.com>