prow, referee: enable pull_request events, fix switch
The referee plugin didn't react to pull_request events, since it was not
enabled to do so in the plugin configuration.

Also switch in go works differently - there's no fallthru. This is fixed
with this change.

Finally we add some more cases to update the metrics, i.e. converted to
draft vs. ready for review.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>