kubemacpool, postsubmit: Fix to push last added tag
Currently the postsubmit takes the first tag that is set on HEAD.
This assumes that there is only one tag.
However, since this postsubmit is triggered by the tag added, it only
makes sense the it should take the last one.
This change would fix issues where a new release tag is set on a commit
with an already set tag (theoretically valid). What will happen then is
that the new tag will be taken, and not the old one.

Signed-off-by: Ram Lavi <ralavi@redhat.com>