kubevirt, labels: rename code-quality into cleanup
The label `sig/code-quality` is used to mark pull requests that target
improvements to our code base. However, the sig prefix gives the
impression that there is a sig that is actively caring for such PRs.

Given the fact that there's neither an entry inside sigs.yaml, nor a
proper sig folder in the community repository, nor any publicly held
meetings that are announced in the community calendar, we can assume
that such a sig doesn't exist.

During the discussion at the community meeting about this topic there
was the suggestion to just rename the label into a `kind` label. Looking
at Kubernetes labels provided a general label `kind/cleanup`[1] that
matches a broader set of scenarios and seems to be well suited for
covering all the code-quality PRs.

Therefore this PR renames the label `sig/code-quality` to
`kind/cleanup`.

[1]: https://github.com/kubernetes/test-infra/blob/master/label_sync/labels.md#kind/cleanup

Signed-off-by: Daniel Hiller <dhiller@redhat.com>