approvers: rename label from needs/approver-review
Since the initial approach (using the `/needs ...`) is not supported
by the label plugin, we use the `/label` command. This requires a
configuration in the label plugin section to allow the bot user to set
the label on PRs, which we add here.

Finally we update the labels.md page with the latest changes.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>