approvers: add commenter that marks stale PRs (#3655)
* approvers: add commenter that marks stale PRs

Adds a periodic job using the commenter to mark PRs that haven't
received a review by an approver within a week with the
needs/approver-review label.

Since that label doesn't exist yet, it is added as well - furthermore
the documentation page is updated.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* approvers: rename label from needs/approver-review

Since the initial approach (using the `/needs ...`) is not supported
by the label plugin, we use the `/label` command. This requires a
configuration in the label plugin section to allow the bot user to set
the label on PRs, which we add here.

Finally we update the labels.md page with the latest changes.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>